+++ src/backend/common/interfaces/IRxStorage.h
@@ -45,6 +45,7 @@ class IRxStorage
 public:
     virtual ~IRxStorage() = default;
 
+    virtual bool isAllocated() const                                                                                            = 0;
     virtual HugePagesInfo hugePages() const                                                                                     = 0;
     virtual RxDataset *dataset(const Job &job, uint32_t nodeId) const                                                           = 0;
     virtual void init(const RxSeed &seed, uint32_t threads, bool hugePages, bool oneGbPages, RxConfig::Mode mode, int priority) = 0;