+++ src/backend/cpu/interfaces/ICpuInfo.h
@@ -37,10 +37,18 @@ namespace xmrig {
 class ICpuInfo
 {
 public:
-    enum Vendor {
+    enum Vendor : uint32_t {
         VENDOR_UNKNOWN,
         VENDOR_INTEL,
         VENDOR_AMD
+    };
+
+    enum MsrMod : uint32_t {
+        MSR_MOD_NONE,
+        MSR_MOD_RYZEN,
+        MSR_MOD_INTEL,
+        MSR_MOD_CUSTOM,
+        MSR_MOD_MAX
     };
 
     virtual ~ICpuInfo() = default;
@@ -58,6 +66,7 @@ public:
     virtual const char *backend() const                                             = 0;
     virtual const char *brand() const                                               = 0;
     virtual CpuThreads threads(const Algorithm &algorithm, uint32_t limit) const    = 0;
+    virtual MsrMod msrMod() const                                                   = 0;
     virtual size_t cores() const                                                    = 0;
     virtual size_t L2() const                                                       = 0;
     virtual size_t L3() const                                                       = 0;