+++ src/base/kernel/Base.cpp
@@ -127,7 +127,7 @@ private:
             return config.release();
         }
 
-        chain.addFile(process->location(Process::ExeLocation, "config.json"));
+        chain.addFile(Process::location(Process::ExeLocation, "config.json"));
 
         if (read(chain, config)) {
             return config.release();
@@ -177,7 +177,7 @@ int xmrig::Base::init()
     Platform::init(config()->userAgent());
 
     if (isBackground()) {
-        Log::background = true;
+        Log::setBackground(true);
     }
     else {
         Log::add(new ConsoleLog());