+++ src/core/config/usage.h
@@ -89,7 +89,8 @@ static inline const std::string &usage()
     u += "      --randomx-no-numa         disable NUMA support for RandomX\n";
     u += "      --randomx-mode=MODE       RandomX mode: auto, fast, light\n";
     u += "      --randomx-1gb-pages       use 1GB hugepages for dataset (Linux only)\n";
-    u += "      --randomx-wrmsr=N         write value (0-15) to Intel MSR register 0x1a4 or do nothing (-1) (Linux only)\n";
+    u += "      --randomx-wrmsr=N         write custom value (0-15) to Intel MSR register 0x1a4 or disable MSR mod (-1)\n";
+    u += "      --randomx-no-rdmsr        disable reverting initial MSR values on exit\n";
 #   endif
 
 #   ifdef XMRIG_FEATURE_HTTP
@@ -136,6 +137,7 @@ static inline const std::string &usage()
     u += "      --health-print-time=N     print health report every N seconds\n";
 #   endif
     u += "      --no-color                disable colored output\n";
+    u += "      --verbose                 verbose output\n";
 
     u += "\nMisc:\n";