+++ src/crypto/randomx/jit_compiler_x86.hpp
@@ -69,10 +69,14 @@ namespace randomx {
 		int registerUsage[RegistersCount];
 		uint8_t* allocatedCode;
 		uint8_t* code;
+#		ifdef XMRIG_FIX_RYZEN
+		std::pair<const void*, const void*> mainLoopBounds;
+#		endif
 		int32_t codePos;
 		uint32_t vm_flags;
 
 		static bool BranchesWithin32B;
+		bool hasAVX;
 
 		static void applyTweaks();
 		void generateProgramPrologue(Program&, ProgramConfiguration&);