+++ src/crypto/rx/RxBasicStorage.h
@@ -48,6 +48,7 @@ public:
     ~RxBasicStorage() override;
 
 protected:
+    bool isAllocated() const override;
     HugePagesInfo hugePages() const override;
     RxDataset *dataset(const Job &job, uint32_t nodeId) const override;
     void init(const RxSeed &seed, uint32_t threads, bool hugePages, bool oneGbPages, RxConfig::Mode mode, int priority) override;